fix: use same image crossOrigin in preloading and img tag
The <img> HTML tag already supports the `crossOrigin` property, and Chakra's `<Image>` will already forward the `crossOrigin` prop to the tag, hooray!

But, when we _preload_ the image, we do so without `crossOrigin` set, which means that the preloaded image can't actually be reused for the final `<img>` tag 😓 This means that the browser sends a second request, and preloading is not only unhelpful, but more expensive on bandwidth and delays the start of the actual final image request!

In this change, we update the `<Image>` component to use the `crossOrigin` prop in its image preload request, if present. Now, when an `<Image>` is provided a `crossOrigin` prop, we'll make only one request, with the `crossOrigin` behavior that the caller requested.