fix(transition): pass nodeRef to React transition group to avoid warning
This change passes a reference to the Transition component, which silences a warning. This means
that the API of our Transition component has to change as well!

BREAKING CHANGE: The Transition, Slide, Fade, Scale and SlideFade components now returns an object
with a ref prop and a style prop instead of just the style object. This means you now have to spread
the children argument on the element that is being transitioned.

fix #1496