Add useDisclosure to imports
Many of the examples shown make use of the `useDisclosure` hook. If we are expecting a new developer to be able to copy and paste the code displayed here and have this *just work* then I think it makes sense to add the `useDisclosure` hook to the import example otherwise the developer needs to read up on where that hook can be imported from.