fix(use-outside-click): register clicks on document
I expect useOutsideClick() to respond to clicks anywhere on the page,
but it was ignoring clicks outside <body>. There can be space below the
end of the body for <body> elements who's height is less than the
browsers height.

Now, we respond to events on document too. the test added in the prior
commit now passes.

Fixes #5837