fix: use-outside-click register clicks on document (#5838)
* test: add tests for use-outside-click

It's currently failing because clicks on the document aren't registered.
We want this because sometimes the body is shorter than the browser
height and a click in the blank space outside of the body isn't
registered.

* fix(use-outside-click): register clicks on document

I expect useOutsideClick() to respond to clicks anywhere on the page,
but it was ignoring clicks outside <body>. There can be space below the
end of the body for <body> elements who's height is less than the
browsers height.

Now, we respond to events on document too. the test added in the prior
commit now passes.

Fixes #5837

* test: add types in the use-outside-click test

Thanks Typescript!